python3 compatibility

+++ tools.py
@@ -32,7 +32,7 @@ def args_are_valid(args, names, intervals):
 
 def get_res_string(res):
 	"""Converts resolution in bp to string (e.g. 10kb)"""
-	res_kb = res/1000
+	res_kb = int(res/1000)
 	if res_kb < 1000:
 		return str(res_kb) + "kb"
 	else:
